neuralprophet/requirements.txt
requests>=2.31.0
pyyaml>=6.0.1
pandas>=2.0.0
neuralprophet>=0.7.0